Fix from PR
