+++ front/src/modules/metadata/hooks/useCreateOneCustomObject.ts
@@ -1,0 +1,8 @@
+// TODO: add zod to validate that we have at least id on each object
+export const useCreateOneCustomObject = ({
+  objectName,
+}: {
+  objectName: string;
+}) => {
+  // TODO : code
+};
